+++ UserInterface/WeekPayrollControl.cs
@@ -92,9 +92,9 @@ namespace UserInterface
         {
             int selectedRow = Int32.Parse(mg_weeks.SelectedRows[0].Index.ToString());
             int codeEmployee = Int32.Parse(mg_weeks.Rows[selectedRow].Cells[0].Value.ToString());
-            int totalHours = Int32.Parse(mg_weeks.Rows[selectedRow].Cells[2].Value.ToString());
-            int ordinaryHours = Int32.Parse(mg_weeks.Rows[selectedRow].Cells[3].Value.ToString());
-            int extraHours = Int32.Parse(mg_weeks.Rows[selectedRow].Cells[4].Value.ToString());
+            Double totalHours = Double.Parse(mg_weeks.Rows[selectedRow].Cells[2].Value.ToString());
+            Double ordinaryHours = Double.Parse(mg_weeks.Rows[selectedRow].Cells[3].Value.ToString());
+            Double extraHours = Double.Parse(mg_weeks.Rows[selectedRow].Cells[4].Value.ToString());
             int weekDBCode = Int32.Parse(mg_weeks.Rows[selectedRow].Cells[5].Value.ToString());
 
             tb_filter.Text = "";
